<WIP> fix call instruction on stack
Special case on MAC M1, the EXC_BAD_ACC (also shown as a segment fault)
on the entry of code on stack is uncatchable. Therefore, the exceptional
exit status (256 + TERM Signal No.) is used as an acceptable results.
Also a bug in the run-test scheduler is fixed to collect the actual
exception signal number.