<WIP> fix return-without-call on MAC M1
It seems that the stack space is protected. When faking a frame by
pushing RA on stack and making SP out of the previous frame space (just
by more than 8 bytes) triggers a BUS access error!

We would need further tests on the access (read/write) range from SP on
MAC M1 to test this type of protection.